% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.FitARMA.R
\name{coef.FitARMA}
\alias{coef.FitARMA}
\title{coef method for class FitARMA}
\usage{
coef(object, ...) ## S3 method for class 'FitARMA'
}
\arguments{
\item{object}{class FitARMA object.}

\item{...}{auxiliary parameters.}
}
\value{
matrix with 3 columns.
}
\description{
Produces table showing parameters, standard errors and Z-ratios.
}
\examples{
data(SeriesA)
out<-FitARMA(SeriesA, c(1,0,1))
coef(out)

}
\seealso{
\code{\link[=FitARMA]{FitARMA()}}.
}
\author{
A.I. McLeod.
}
