% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.FitARMA.R
\name{fitted.FitARMA}
\alias{fitted.FitARMA}
\title{fitted method for class FitARMA}
\usage{
fitted(object, ...) ## S3 method for class 'FitARMA'
}
\arguments{
\item{object}{class FitARMA object.}

\item{...}{auxiliary parameters.}
}
\value{
vector or ts object.
}
\description{
The fitted values are the observed minus residuals. If there is differencing,
the observed values are those corresponding to the differenced time series.
}
\examples{
data(SeriesA)
out<-FitARMA(SeriesA, c(1,0,1))
fitted(out)

}
\seealso{
\code{\link[=FitARMA]{FitARMA()}}.
}
\author{
A.I. McLeod.
}
