% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitARMA.R
\name{GetFitARMA}
\alias{GetFitARMA}
\title{Fit ARMA(p,q) model to mean zero time series.}
\usage{
GetFitARMA(y, p, q, pApprox = 30, init = 0)
}
\arguments{
\item{y}{time series.}

\item{p}{AR order.}

\item{q}{MA order.}

\item{pApprox}{AR approximation.}

\item{init}{initial parameter estimates.}
}
\value{
A list with class name "FitARMA" and components:
\itemize{
\item \code{loglikelihood} value of maximized loglikelihood.
\item \code{phiHat} estimated phi parameters.
\item \code{thetaHat} estimated theta parameters.
\item \code{convergence} result from optim.
\item \code{algorithm} indicates "L-BFGS-B" or "Nelder-Mead" according as which algorithm was used in optim.
}
}
\description{
The algorithm of McLeod and Zhang (2007) is used.
}
\details{
See McLeod and Zhang (2006).
}
\examples{
data(SeriesA)
z<-SeriesA-mean(SeriesA)
GetFitARMA(z, 1, 1)
w<-diff(z, differences=1)
GetFitARMA(w, 0, 1) 

}
\references{
A.I. McLeod andY. Zhang (2008), Faster ARMA maximum likelihood estimation,
Computational Statistics & Data Analysis, 52-4, 2166-2176. DOI link:
http://dx.doi.org/10.1016/j.csda.2007.07.020.
}
\seealso{
\code{\link[=arima]{arima()}}, \code{\link[=FitARMA]{FitARMA()}}.
}
\author{
A.I. McLeod.
}
