% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpulseCoefficientsARMA.R
\name{ImpulseCoefficientsARMA}
\alias{ImpulseCoefficientsARMA}
\title{Impulse coefficients of ARMA}
\usage{
ImpulseCoefficientsARMA(phi, theta, lag.max)
}
\arguments{
\item{phi}{AR coefficients.}

\item{theta}{MA coefficients.}

\item{lag.max}{lags 0,...,lag.max.}
}
\value{
vector length lag.max+1.
}
\description{
The coefficients in the infinite MA expansion of the ARMA model are determined.
}
\examples{
ImpulseCoefficientsARMA(0.9,0.5,20)

}
\author{
A.I. McLeod.
}
