% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FitARMA.R
\name{print.FitARMA}
\alias{print.FitARMA}
\title{print method for class FitARMA}
\usage{
print(x, ...) ## S3 method for class 'FitARMA'
}
\arguments{
\item{x}{object, class FitARMA.}

\item{...}{optional arguments.}
}
\value{
the result is displayed.
}
\description{
a brief summary is printed out of the fitted model.
}
\examples{
data(SeriesA)
out <- FitARMA(SeriesA, c(1,0,1))
print(out)

}
\seealso{
\code{\link[=FitARMA]{FitARMA()}}.
}
\author{
A.I. McLeod.
}
