% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.FitARMA.R
\name{residuals.FitARMA}
\alias{residuals.FitARMA}
\title{residuals method for class FitARMA}
\usage{
residuals(object, ...) ## S3 method for class 'FitARMA'
}
\arguments{
\item{object}{class FitARMA object.}

\item{...}{auxiliary parameters.}
}
\value{
vector or ts object.
}
\description{
The innovation residuals are obtained.
}
\examples{
data(SeriesA)
out <- FitARMA(SeriesA, c(1,0,1))
resid(out)

}
\seealso{
\code{\link[=FitARMA]{FitARMA()}}.
}
\author{
A.I. McLeod.
}
