% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FitARMA.R
\name{summary.FitARMA}
\alias{summary.FitARMA}
\title{print method for class FitARMA}
\usage{
summary(object, ...) ## S3 method for class 'FitARMA'
}
\arguments{
\item{object}{class FitARMA object.}

\item{...}{optional arguments.}
}
\value{
the result is displayed.
}
\description{
a summary is printed out of the fitted model
}
\examples{
data(SeriesA)
out <- FitARMA(SeriesA, c(1,0,1))
summary(out)

}
\seealso{
\code{\link[=FitARMA]{FitARMA()}}.
}
\author{
A.I. McLeod.
}
