% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccfAR.R
\name{tccfAR}
\alias{tccfAR}
\title{Theoretical cross-covariances of auxilary AR process in ARMA(p,q)}
\usage{
tccfAR(phi, theta)
}
\arguments{
\item{phi}{AR coefficients in ARMA.}

\item{theta}{MA coefficients in ARMA.}
}
\value{
vector of cross-covariances.
}
\description{
The auxilary AR processes in the ARMA(p,q) model phi(B)z(t)=theta(B)a(t) are
defined by phi(B)u(t)=-a(t) and theta(B)v(t)=a(t). The upper off-diagonal p-by-q
block of the ARMA information matrix is obtained from the cross-covariances
of u(t) and v(t). This function obtains these covariances.
}
\details{
A set of linear equations which determine the covariances is solved.
The algorithm is similar in spirit to that for the autocovariances (McLeod, 1975).
}
\examples{
tccfAR(0.9,0.5)

}
\references{
McLeod, A.I. (1975), Derivation of the theoretical
autocorrelation function of autoregressive moving-average time
series, Applied Statistics 24, 255-256.
}
\seealso{
\code{\link[=InformationMatrixARMA]{InformationMatrixARMA()}}.
}
\author{
A.I. McLeod.
}
